# packages/saas-platform-spec/admin-api.openapi.yaml
openapi: 3.1.0
...
info:
    title: SaaS-Plattform SuperAdmin-API
    version: 0.1.0-draft
    description: |
        Read- und Write-Operationen für die Plattform-Verwaltung:
        Tenants, Users, Subscriptions, Promo-Codes, Audit-Log,
        Manifest-Discovery, plus Plan-Catalog-Lookup.
        Authentifizierung via JWT mit Rolle SUPER_ADMIN; alle schreibenden
        Operationen verlangen zusätzlich einen TOTP-Code im Header
        X-Mfa-Code, sofern der Admin MFA aktiviert hat (sonst 401
        MFA_REQUIRED). Manifest ist nie Security-Quelle (siehe SPEC §10.4).
servers:
    - url: https://{host}/api/v1/admin
      variables:
          host:
              default: localhost:3000
              description: Konsumenten-spezifischer Host (Domain der jeweiligen App)
...
security:
    - bearerAuth: []
paths:
...
    /catalog/business-types/{id}:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: BusinessType-Stamm Detail
            responses: { '200': { description: Detail } }
        patch:
            tags: [catalog]
            summary: BusinessType-Stamm aktualisieren
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogStampUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
        delete:
            tags: [catalog]
            summary: BusinessType-Stamm soft-deleten
            x-mfa-required: true
            responses: { '200': { description: Gelöscht } }
    /catalog/business-types/{id}/versions:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: BusinessTypeVersionen auflisten
            responses: { '200': { description: Liste } }
        post:
            tags: [catalog]
            summary: BusinessTypeVersion-Entwurf anlegen (Bundle-Zusammenstellung)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogBusinessTypeVersionDraft' } } }
            responses: { '201': { description: Entwurf angelegt } }
...
    /catalog/marketing-projections/{id}:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: MarketingProjection Detail
            responses: { '200': { description: Detail } }
        patch:
            tags: [catalog]
            summary: MarketingProjection aktualisieren
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogMarketingProjectionUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
        delete:
            tags: [catalog]
            summary: MarketingProjection löschen
            x-mfa-required: true
            responses: { '200': { description: Gelöscht } }
...
    /catalog/features/{key}:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Feature-Katalog-Eintrag pflegen (Label/Beschreibung/Icon/Tier)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryBaseUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
    /catalog/features/{key}/review:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Feature-Freigabe-Status setzen
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryReview' } } }
            responses: { '200': { description: Gesetzt } }
    /catalog/features/{key}/i18n:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Feature-i18n setzen (Locale-Übersetzungen)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryI18nUpdate' } } }
            responses: { '200': { description: Gesetzt } }
    /catalog/quotas/{key}:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Quota-Katalog-Eintrag pflegen (Label/Beschreibung)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryBaseUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
    /catalog/quotas/{key}/review:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Quota-Freigabe-Status setzen
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryReview' } } }
            responses: { '200': { description: Gesetzt } }
    /catalog/quotas/{key}/i18n:
        parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
        patch:
            tags: [catalog]
            summary: Quota-i18n setzen
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogEntryI18nUpdate' } } }
            responses: { '200': { description: Gesetzt } }
    /catalog/discovery/sync:
        post:
            tags: [catalog]
            summary: Discovery-Snapshot in den Katalog synchronisieren
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogDiscoverySync' } } }
            responses: { '200': { description: Synchronisiert } }

    # ── Discovery (Scan/Status, außerhalb /catalog) ──
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
    schemas:
...
        CatalogStampUpdate:
            type: object
            description: 'Stamm-Update für Plan ODER BusinessType'
            properties:
                label: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 2000, nullable: true }
                icon: { type: string, maxLength: 64, nullable: true }
                sortOrder: { type: integer, minimum: 0, maximum: 10000 }
...
        CatalogMarketingProjectionUpdate:
            type: object
            description: 'Partial-Update einer MarketingProjection'
            properties:
                displayLabel: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, minLength: 1, maxLength: 2000 }
                visible: { type: boolean }
                badge: { type: string, maxLength: 40 }
                topFeatures:
                    type: array
                    items:
                        type: object
                        required: [label, strong]
                        properties:
                            key: { type: string, maxLength: 120 }
                            label: { type: string, maxLength: 120 }
                            strong: { type: string, maxLength: 80 }
                trialEnabled: { type: boolean }
                trialDays: { type: integer, minimum: 1, maximum: 365 }
                priceTag: { type: string, maxLength: 120, nullable: true }
                ctaLabel: { type: string, maxLength: 120, nullable: true }
                priority: { type: integer, minimum: 0, maximum: 10000 }
                highlight: { type: boolean }
...
        CatalogBusinessTypeVersionDraft:
            type: object
            required: [bundles]
            properties:
                bundles:
                    type: array
                    minItems: 1
                    items:
                        type: object
                        required: [bundleVersionId]
                        properties:
                            bundleVersionId: { type: string, format: uuid }
                            sortOrder: { type: integer, minimum: 0, maximum: 10000 }
                quotaOverrides: { type: object, additionalProperties: { type: number } }
                monthlyNet: { type: string, pattern: '^\d+(\.\d{1,2})?$', nullable: true }
                yearlyNet: { type: string, pattern: '^\d+(\.\d{1,2})?$', nullable: true }
                marketed: { type: boolean }
                changeNote: { type: string, maxLength: 2000 }
                baseVersionId: { type: string, format: uuid, nullable: true }
...
        CatalogEntryReview:
            type: object
            required: [discoveryStatus]
            properties:
                discoveryStatus: { type: string, enum: [pending, approved, outdated, obsolete] }
        CatalogEntryI18nUpdate:
            type: object
            required: [i18n]
            properties:
                i18n:
                    type: object
                    additionalProperties:
                        type: object
                        properties:
                            label: { type: string }
                            description: { type: string }
                    description: 'Locale-Map → {label?, description?}'
        CatalogEntryBaseUpdate:
            type: object
            properties:
                label: { type: string, maxLength: 120 }
                description: { type: string, maxLength: 2000, nullable: true }
                icon: { type: string, maxLength: 64, nullable: true, description: 'nur Feature' }
                tier: { type: string, maxLength: 40, nullable: true, description: 'nur Feature, logische Gruppe' }
        CatalogDiscoverySync:
            type: object
            required: [snapshot]
            properties:
                snapshot:
                    type: object
                    additionalProperties: true
                    description: 'DiscoverySnapshot (app/features/quotas/capabilities) aus @saasicat/types'
...